requirements-minimal.txt
# Core dependencies (minimal - no NER)
pymongo>=4.6.0
pydantic>=2.5.0
python-Levenshtein>=0.23.0
jellyfish>=1.0.0
duckdb>=0.10.0
pandas>=2.1.0
click>=8.1.0
tqdm>=4.66.0
pyyaml>=6.0.0

# Development dependencies
pytest>=7.4.0
pytest-benchmark>=4.0.0
mypy>=1.7.0
black>=23.12.0
ruff>=0.1.0
